+++ .gitignore
@@ -4,5 +4,5 @@
 erl_crash.dump
 /config
 *.ez
-clash_of_clans_slackbot
+/clash_of_clans_slackbot
 *.beam
