+++ lib/clash_of_clans_slackbot/repositories/clashcaller.ex
