+++ lib/clash_of_clans_slackbot/validator.ex
@@ -1,0 +1,10 @@
+defmodule Validator do
+  def validate_token(token) do
+    case !!token do
+      true -> { :ok, token }
+      _    -> { :err, "No token provided" }
+    end
+  end
+end
+
+
